+++ accounts-db/src/accounts_db/tests/impl.rs
@@ -134,7 +134,7 @@ fn create_store_for_shrink_tests(
     ));
     accounts_db.storage.insert(Arc::clone(&store));
     store.add_accounts(num_tombstones.max(1), alive_bytes);
-    store.batch_insert_tombstone_offsets(0..num_tombstones);
+    store.batch_insert_tombstone_offsets(0..num_tombstones as u64);
     (temp_dir, store)
 }
